Added our Certificate of Incorporation to the website.
This will be used in determining our status as time goes on and for getting the resources we need. The certificate of Incorporation is just like an operating certificate or anything like that. There is no information on that isn't public. It's just a place of reference for businesses that we may be doing business with.